+++ model/@stk_model_/stk_predict_.m
@@ -1,0 +1,36 @@
+% STK_PREDICT_ [STK internal]
+%
+% See also: stk_predict
+
+% Copyright Notice
+%
+%    Copyright (C) 2020 CentraleSupelec
+%
+%    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
+
+% Copying Permission Statement
+%
+%    This file is part of
+%
+%            STK: a Small (Matlab/Octave) Toolbox for Kriging
+%               (http://sourceforge.net/projects/kriging)
+%
+%    STK is free software: you can redistribute it and/or modify it under
+%    the terms of the GNU General Public License as published by the Free
+%    Software Foundation,  either version 3  of the License, or  (at your
+%    option) any later version.
+%
+%    STK is distributed  in the hope that it will  be useful, but WITHOUT
+%    ANY WARRANTY;  without even the implied  warranty of MERCHANTABILITY
+%    or FITNESS  FOR A  PARTICULAR PURPOSE.  See  the GNU  General Public
+%    License for more details.
+%
+%    You should  have received a copy  of the GNU  General Public License
+%    along with STK.  If not, see <http://www.gnu.org/licenses/>.
+
+function [zp, lambda, mu, K] = stk_predict_ (model, xt)  %#ok<STOUT,INUSD>
+
+stk_error (sprintf (['stk_predict_ is not implemented for models of ' ...
+    'class %s.'], class (model)), 'NotImplemented');
+
+end % function