+++ admin/build_tools/build_octpkg.m
@@ -110,7 +110,7 @@ fprintf (fid, 'License: GPLv3+\n');
 fprintf (fid, '#\n');
 fprintf (fid, 'Url: https://github.com/stk-kriging/stk/\n');
 fprintf (fid, '#\n');
-fprintf (fid, 'Depends: octave (>= 3.8.0)\n');
+fprintf (fid, 'Depends: octave (>= 4.0.0)\n');
 fprintf (fid, '#\n');
 fprintf (fid, 'Autoload: no\n');
 fclose (fid);